@@ -4,7 +4,7 @@
   "description": "HAP-NodeJS is a Node.js implementation of HomeKit Accessory Server.",
   "main": "index.js",
   "engines": {
-    "node": ">=4.3.2"
+    "node": ">=5.10.0"
   },
   "dependencies": {
     "curve25519-n": "^1.1.0",
@@ -13,7 +13,7 @@
     "ip": "^1.1.3",
     "mdns": "^2.3.2",
     "node-persist": "^0.0.8",
-    "srp": "git://github.com/KhaosT/node-srp"
+    "fast-srp-hap": "^1.0.0"
   },
   "devDependencies": {
     "simple-plist": "0.0.4"
