@@ -1,6 +1,6 @@
 var debug = require('debug')('HAPServer');
 var crypto = require("crypto");
-var srp = require("srp");
+var srp = require("fast-srp-hap");
 var url = require("url");
 var inherits = require('util').inherits;
 var ed25519 = require("ed25519");
@@ -190,7 +190,7 @@ HAPServer.prototype._onRequest = function(request, response, session, events) {
   debug("[%s] HAP Request: %s %s", this.accessoryInfo.username, request.method, request.url);
   
   // collect request data, if any
-  var requestData = Buffer("");
+  var requestData = Buffer.alloc(0);
   request.on('data', function(data) { requestData = Buffer.concat([requestData, data]); });
   request.on('end', function() {
     
@@ -319,7 +319,7 @@ HAPServer.prototype._handlePairStepOne = function(request, response, session) {
   srp.genKey(32, function (error, key) {
     
     // create a new SRP server
-    var srpServer = new srp.Server(srpParams, Buffer(salt), Buffer("Pair-Setup"), Buffer(this.accessoryInfo.pincode), key);
+    var srpServer = new srp.Server(srpParams, Buffer.from(salt), Buffer.from("Pair-Setup"), Buffer.from(this.accessoryInfo.pincode), key);
     var srpB = srpServer.computeB();
 
     // attach it to the current TCP session
@@ -381,20 +381,20 @@ HAPServer.prototype._handlePairStepThree = function(request, response, session, 
 
   var encryptedData = objects[HAPServer.Types.ENCRYPTED_DATA];
   
-  var messageData = Buffer(encryptedData.length - 16);
-  var authTagData = Buffer(16);
+  var messageData = Buffer.alloc(encryptedData.length - 16);
+  var authTagData = Buffer.alloc(16);
   
   encryptedData.copy(messageData, 0, 0, encryptedData.length - 16);
   encryptedData.copy(authTagData, 0, encryptedData.length - 16, encryptedData.length);
 
   var S_private = srpServer.computeK();
-  var encSalt = Buffer("Pair-Setup-Encrypt-Salt");
-  var encInfo = Buffer("Pair-Setup-Encrypt-Info");
+  var encSalt = Buffer.from("Pair-Setup-Encrypt-Salt");
+  var encInfo = Buffer.from("Pair-Setup-Encrypt-Info");
 
   var outputKey = hkdf.HKDF("sha512", encSalt, S_private, encInfo, 32);
   
-  var plaintextBuffer = Buffer(messageData.length);
-  encryption.verifyAndDecrypt(outputKey, Buffer("PS-Msg05"), messageData, authTagData, null, plaintextBuffer);
+  var plaintextBuffer = Buffer.alloc(messageData.length);
+  encryption.verifyAndDecrypt(outputKey, Buffer.from("PS-Msg05"), messageData, authTagData, null, plaintextBuffer);
   
   // decode the client payload and pass it on to the next step
   var M5Packet = tlv.decode(plaintextBuffer);
@@ -412,8 +412,8 @@ HAPServer.prototype._handlePairStepFour = function(request, response, session, c
   debug("[%s] Pair step 4/5", this.accessoryInfo.username);
 
   var S_private = session.srpServer.computeK();
-  var controllerSalt = Buffer("Pair-Setup-Controller-Sign-Salt");
-  var controllerInfo = Buffer("Pair-Setup-Controller-Sign-Info");
+  var controllerSalt = Buffer.from("Pair-Setup-Controller-Sign-Salt");
+  var controllerInfo = Buffer.from("Pair-Setup-Controller-Sign-Info");
   var outputKey = hkdf.HKDF("sha512", controllerSalt, S_private, controllerInfo, 32);
   
   var completeData = Buffer.concat([outputKey, clientUsername, clientLTPK]);
@@ -438,15 +438,15 @@ HAPServer.prototype._handlePairStepFive = function(request, response, session, c
   debug("[%s] Pair step 5/5", this.accessoryInfo.username);
   
   var S_private = session.srpServer.computeK();
-  var accessorySalt = Buffer("Pair-Setup-Accessory-Sign-Salt");
-  var accessoryInfo = Buffer("Pair-Setup-Accessory-Sign-Info");
+  var accessorySalt = Buffer.from("Pair-Setup-Accessory-Sign-Salt");
+  var accessoryInfo = Buffer.from("Pair-Setup-Accessory-Sign-Info");
   var outputKey = hkdf.HKDF("sha512", accessorySalt, S_private, accessoryInfo, 32);
 
   var serverLTPK = this.accessoryInfo.signPk;
-  var usernameData = Buffer(this.accessoryInfo.username);
+  var usernameData = Buffer.from(this.accessoryInfo.username);
   
   var material = Buffer.concat([outputKey, usernameData, serverLTPK]);
-  var privateKey = Buffer(this.accessoryInfo.signSk);
+  var privateKey = Buffer.from(this.accessoryInfo.signSk);
 
   var serverProof = ed25519.Sign(material, privateKey);
 
@@ -456,9 +456,9 @@ HAPServer.prototype._handlePairStepFive = function(request, response, session, c
     HAPServer.Types.PROOF, serverProof
   );
   
-  var ciphertextBuffer = Buffer(Array(message.length));
-  var macBuffer = Buffer(Array(16));
-  encryption.encryptAndSeal(hkdfEncKey, Buffer("PS-Msg06"), message, null, ciphertextBuffer, macBuffer);
+  var ciphertextBuffer = Buffer.alloc(message.length);
+  var macBuffer = Buffer.alloc(16);
+  encryption.encryptAndSeal(hkdfEncKey, Buffer.from("PS-Msg06"), message, null, ciphertextBuffer, macBuffer);
 
   // finally, notify listeners that we have been paired with a client
   this.emit('pair', clientUsername.toString(), clientLTPK, once(function(err) {
@@ -511,13 +511,13 @@ HAPServer.prototype._handlePairVerifyStepOne = function(request, response, sessi
   var publicKey = encryption.generateCurve25519PublicKeyFromSecretKey(secretKey);
   var sharedSec = encryption.generateCurve25519SharedSecKey(secretKey, clientPublicKey);
     
-  var usernameData = Buffer(this.accessoryInfo.username);
+  var usernameData = Buffer.from(this.accessoryInfo.username);
   var material = Buffer.concat([publicKey, usernameData, clientPublicKey]);
-  var privateKey = Buffer(this.accessoryInfo.signSk);
+  var privateKey = Buffer.from(this.accessoryInfo.signSk);
   var serverProof = ed25519.Sign(material, privateKey);
 
-  var encSalt = Buffer("Pair-Verify-Encrypt-Salt");
-  var encInfo = Buffer("Pair-Verify-Encrypt-Info");
+  var encSalt = Buffer.from("Pair-Verify-Encrypt-Salt");
+  var encInfo = Buffer.from("Pair-Verify-Encrypt-Info");
 
   var outputKey = hkdf.HKDF("sha512", encSalt, sharedSec, encInfo, 32).slice(0,32);
 
@@ -539,9 +539,9 @@ HAPServer.prototype._handlePairVerifyStepOne = function(request, response, sessi
   );
   
   // encrypt the response
-  var ciphertextBuffer = Buffer(Array(message.length));
-  var macBuffer = Buffer(Array(16));
-  encryption.encryptAndSeal(outputKey, Buffer("PV-Msg02"), message, null, ciphertextBuffer, macBuffer);
+  var ciphertextBuffer = Buffer.alloc(message.length);
+  var macBuffer = Buffer.alloc(16);
+  encryption.encryptAndSeal(outputKey, Buffer.from("PV-Msg02"), message, null, ciphertextBuffer, macBuffer);
 
   response.writeHead(200, {"Content-Type": "application/pairing+tlv8"});
   response.end(tlv.encode(
@@ -556,17 +556,17 @@ HAPServer.prototype._handlePairVerifyStepTwo = function(request, response, sessi
 
   var encryptedData = objects[HAPServer.Types.ENCRYPTED_DATA];
   
-  var messageData = Buffer(encryptedData.length - 16);
-  var authTagData = Buffer(16);
+  var messageData = Buffer.alloc(encryptedData.length - 16);
+  var authTagData = Buffer.alloc(16);
   encryptedData.copy(messageData, 0, 0, encryptedData.length - 16);
   encryptedData.copy(authTagData, 0, encryptedData.length - 16, encryptedData.length);
 
-  var plaintextBuffer = Buffer(messageData.length);
+  var plaintextBuffer = Buffer.alloc(messageData.length);
   
   // instance of HAPEncryption (created in handlePairVerifyStepOne)
   var enc = session.encryption;
   
-  if (!encryption.verifyAndDecrypt(enc.hkdfPairEncKey, Buffer("PV-Msg03"), messageData, authTagData, null, plaintextBuffer)) {
+  if (!encryption.verifyAndDecrypt(enc.hkdfPairEncKey, Buffer.from("PV-Msg03"), messageData, authTagData, null, plaintextBuffer)) {
     debug("[%s] M3: Invalid signature", this.accessoryInfo.username);
     
     response.writeHead(200, {"Content-Type": "application/pairing+tlv8"});
@@ -621,9 +621,9 @@ HAPServer.prototype._handlePairVerifyStepTwo = function(request, response, sessi
   // TCP-level encryption. We'll do this by adding some more encryption vars to the session, and using them
   // in future calls to onEncrypt, onDecrypt.
   
-  var encSalt = new Buffer("Control-Salt");
-  var infoRead = new Buffer("Control-Read-Encryption-Key");
-  var infoWrite = new Buffer("Control-Write-Encryption-Key");
+  var encSalt = Buffer.from("Control-Salt");
+  var infoRead = Buffer.from("Control-Read-Encryption-Key");
+  var infoWrite = Buffer.from("Control-Write-Encryption-Key");
 
   enc.accessoryToControllerKey = hkdf.HKDF("sha512", encSalt, enc.sharedSec, infoRead, 32);
   enc.controllerToAccessoryKey = hkdf.HKDF("sha512", encSalt, enc.sharedSec, infoWrite, 32);
@@ -653,13 +653,13 @@ HAPServer.prototype._handleRemotePairVerifyStepOne = function(request, remoteSes
   var publicKey = encryption.generateCurve25519PublicKeyFromSecretKey(secretKey);
   var sharedSec = encryption.generateCurve25519SharedSecKey(secretKey, clientPublicKey);
     
-  var usernameData = Buffer(this.accessoryInfo.username);
+  var usernameData = Buffer.from(this.accessoryInfo.username);
   var material = Buffer.concat([publicKey, usernameData, clientPublicKey]);
-  var privateKey = Buffer(this.accessoryInfo.signSk);
+  var privateKey = Buffer.from(this.accessoryInfo.signSk);
   var serverProof = ed25519.Sign(material, privateKey);
 
-  var encSalt = Buffer("Pair-Verify-Encrypt-Salt");
-  var encInfo = Buffer("Pair-Verify-Encrypt-Info");
+  var encSalt = Buffer.from("Pair-Verify-Encrypt-Salt");
+  var encInfo = Buffer.from("Pair-Verify-Encrypt-Info");
 
   var outputKey = hkdf.HKDF("sha512", encSalt, sharedSec, encInfo, 32).slice(0,32);
 
@@ -681,9 +681,9 @@ HAPServer.prototype._handleRemotePairVerifyStepOne = function(request, remoteSes
   );
   
   // encrypt the response
-  var ciphertextBuffer = Buffer(Array(message.length));
-  var macBuffer = Buffer(Array(16));
-  encryption.encryptAndSeal(outputKey, Buffer("PV-Msg02"), message, null, ciphertextBuffer, macBuffer);
+  var ciphertextBuffer = Buffer.alloc(message.length);
+  var macBuffer = Buffer.alloc(16);
+  encryption.encryptAndSeal(outputKey, Buffer.from("PV-Msg02"), message, null, ciphertextBuffer, macBuffer);
 
   var response = tlv.encode(
     HAPServer.Types.SEQUENCE_NUM, 0x02,
@@ -699,17 +699,17 @@ HAPServer.prototype._handleRemotePairVerifyStepTwo = function(request, remoteSes
 
   var encryptedData = objects[HAPServer.Types.ENCRYPTED_DATA];
   
-  var messageData = Buffer(encryptedData.length - 16);
-  var authTagData = Buffer(16);
+  var messageData = Buffer.alloc(encryptedData.length - 16);
+  var authTagData = Buffer.alloc(16);
   encryptedData.copy(messageData, 0, 0, encryptedData.length - 16);
   encryptedData.copy(authTagData, 0, encryptedData.length - 16, encryptedData.length);
 
-  var plaintextBuffer = Buffer(messageData.length);
+  var plaintextBuffer = Buffer.alloc(messageData.length);
   
   // instance of HAPEncryption (created in handlePairVerifyStepOne)
   var enc = session.encryption;
   
-  if (!encryption.verifyAndDecrypt(enc.hkdfPairEncKey, Buffer("PV-Msg03"), messageData, authTagData, null, plaintextBuffer)) {
+  if (!encryption.verifyAndDecrypt(enc.hkdfPairEncKey, Buffer.from("PV-Msg03"), messageData, authTagData, null, plaintextBuffer)) {
     debug("[%s] M3: Invalid signature", this.accessoryInfo.username);
     
     var response = tlv.encode(
@@ -755,9 +755,9 @@ HAPServer.prototype._handleRemotePairVerifyStepTwo = function(request, remoteSes
 
   debug("[%s] Client %s verification complete", this.accessoryInfo.username, clientUsername);
 
-  var encSalt = new Buffer("Control-Salt");
-  var infoRead = new Buffer("Control-Read-Encryption-Key");
-  var infoWrite = new Buffer("Control-Write-Encryption-Key");
+  var encSalt = Buffer.from("Control-Salt");
+  var infoRead = Buffer.from("Control-Read-Encryption-Key");
+  var infoWrite = Buffer.from("Control-Write-Encryption-Key");
 
   enc.accessoryToControllerKey = hkdf.HKDF("sha512", encSalt, enc.sharedSec, infoRead, 32);
   enc.controllerToAccessoryKey = hkdf.HKDF("sha512", encSalt, enc.sharedSec, infoWrite, 32);
@@ -866,7 +866,7 @@ HAPServer.prototype._handleRemoteAccessories = function(request, remoteSession, 
     var response = {
       'configuration-number': this.accessoryInfo.configVersion
     }
-    remoteSession.responseMessage(request, new Buffer(JSON.stringify(response)));
+    remoteSession.responseMessage(request, Buffer.from(JSON.stringify(response)));
   } else {
     var self = this;
     // call out to listeners to retrieve the latest accessories JSON
@@ -881,7 +881,7 @@ HAPServer.prototype._handleRemoteAccessories = function(request, remoteSession, 
         'configuration-number': self.accessoryInfo.configVersion,
         'attribute-database': accessories
       }
-      remoteSession.responseMessage(request, new Buffer(JSON.stringify(response)));
+      remoteSession.responseMessage(request, Buffer.from(JSON.stringify(response)));
     }));
   }
 }
@@ -1063,7 +1063,7 @@ HAPServer.prototype._handleRemoteCharacteristicsWrite = function(request, remote
       }
     }
     
-    remoteSession.responseMessage(request, new Buffer(JSON.stringify(characteristics)));    
+    remoteSession.responseMessage(request, Buffer.from(JSON.stringify(characteristics)));    
   }.bind(this)), true);
 }
 
@@ -1088,7 +1088,7 @@ HAPServer.prototype._handleRemoteCharacteristicsRead = function(request, remoteS
         });
       }
     }
-    remoteSession.responseMessage(request, new Buffer(JSON.stringify(characteristics)));
+    remoteSession.responseMessage(request, Buffer.from(JSON.stringify(characteristics)));
   }.bind(this), true);
 }
 
@@ -1098,13 +1098,13 @@ HAPServer.prototype._handleRemoteCharacteristicsRead = function(request, remoteS
  
 function HAPEncryption() {
   // initialize member vars with null-object values
-  this.clientPublicKey = new Buffer(0);
-  this.secretKey = new Buffer(0);
-  this.publicKey = new Buffer(0);
-  this.sharedSec = new Buffer(0);
-  this.hkdfPairEncKey = new Buffer(0);
+  this.clientPublicKey = Buffer.alloc(0);
+  this.secretKey = Buffer.alloc(0);
+  this.publicKey = Buffer.alloc(0);
+  this.sharedSec = Buffer.alloc(0);
+  this.hkdfPairEncKey = Buffer.alloc(0);
   this.accessoryToControllerCount = { value: 0 };
   this.controllerToAccessoryCount = { value: 0 };
-  this.accessoryToControllerKey = new Buffer(0);
-  this.controllerToAccessoryKey = new Buffer(0);
+  this.accessoryToControllerKey = Buffer.alloc(0);
+  this.controllerToAccessoryKey = Buffer.alloc(0);
 }
