@@ -277,7 +277,7 @@ StreamController.prototype._handleSelectedStreamConfigurationWrite = function(va
   var self = this;
   self.selectedConfiguration = value;
 
-  var data = Buffer(value, 'base64');
+  var data = Buffer.from(value, 'base64');
   var objects = tlv.decode(data);
 
   var session;
@@ -471,7 +471,7 @@ StreamController.prototype._handleStopStream = function(callback, silent) {
 StreamController.prototype._handleSetupWrite = function(value, callback) {
   var self = this;
 
-  var data = Buffer(value, 'base64');
+  var data = Buffer.from(value, 'base64');
   var objects = tlv.decode(data);
 
   self.sessionIdentifier = objects[StreamController.SetupTypes.SESSION_ID];
@@ -593,14 +593,14 @@ StreamController.prototype._generateSetupResponse = function(identifier, respons
 
   var ipVer = 0;
   var ipAddress = null;
-  var videoPort = new Buffer(2);
-  var audioPort = new Buffer(2);
+  var videoPort = Buffer.alloc(2);
+  var audioPort = Buffer.alloc(2);
 
-  var videoSSRC = new Buffer(4);
-  var audioSSRC = new Buffer(4);
+  var videoSSRC = Buffer.alloc(4);
+  var audioSSRC = Buffer.alloc(4);
 
-  var videoSRTP = Buffer([0x01, 0x01, 0x02, 0x02, 0x00, 0x03, 0x00]);
-  var audioSRTP = Buffer([0x01, 0x01, 0x02, 0x02, 0x00, 0x03, 0x00]);
+  var videoSRTP = Buffer.from([0x01, 0x01, 0x02, 0x02, 0x00, 0x03, 0x00]);
+  var audioSRTP = Buffer.from([0x01, 0x01, 0x02, 0x02, 0x00, 0x03, 0x00]);
 
   if (self.requireProxy) {
     let currentAddress = ip.address();
@@ -611,7 +611,7 @@ StreamController.prototype._generateSetupResponse = function(identifier, respons
       ipVer = 0;
     }
 
-    ipAddress = Buffer(currentAddress);
+    ipAddress = Buffer.from(currentAddress);
     videoPort.writeUInt16LE(self.videoProxy.outgoingLocalPort(), 0);
 
     if (!self.disableAudioProxy) {
@@ -769,17 +769,17 @@ StreamController.prototype._supportedVideoStreamConfiguration = function(videoPa
     throw new Error('Video resolutions cannot be undefined');
   }
 
-  var videoAttrsTLV = Buffer(0);
+  var videoAttrsTLV = Buffer.alloc(0);
   resolutions.forEach(function(resolution) {
     if (resolution.length != 3) {
       throw new Error('Unexpected video resolution');
     }
 
-    var imageWidth = new Buffer(2);
+    var imageWidth = Buffer.alloc(2);
     imageWidth.writeUInt16LE(resolution[0], 0);
-    var imageHeight = new Buffer(2);
+    var imageHeight = Buffer.alloc(2);
     imageHeight.writeUInt16LE(resolution[1], 0);
-    var frameRate = new Buffer(1);
+    var frameRate = Buffer.alloc(1);
     frameRate.writeUInt8(resolution[2]);
 
     var videoAttrTLV = tlv.encode(
@@ -817,7 +817,7 @@ StreamController.prototype._supportedAudioStreamConfiguration = function(audioPa
     throw new Error('Audio codecs cannot be undefined');
   }
 
-  var audioConfigurationsBuffer = Buffer(0);
+  var audioConfigurationsBuffer = Buffer.alloc(0);
   var hasSupportedCodec = false;
 
   codecs.forEach(function(codecParam){
