@@ -20,8 +20,8 @@ function AccessoryInfo(username) {
   this.displayName = "";
   this.category = "";
   this.pincode = "";
-  this.signSk = Buffer(0);
-  this.signPk = Buffer(0);
+  this.signSk = Buffer.alloc(0);
+  this.signPk = Buffer.alloc(0);
   this.pairedClients = {}; // pairedClients[clientUsername:string] = clientPublicKey:Buffer
   this.configVersion = 1;
   this.configHash = "";
@@ -106,13 +106,13 @@ AccessoryInfo.load = function(username) {
     info.displayName = saved.displayName || "";
     info.category = saved.category || "";
     info.pincode = saved.pincode || "";
-    info.signSk = new Buffer(saved.signSk || '', 'hex');
-    info.signPk = new Buffer(saved.signPk || '', 'hex');
+    info.signSk = Buffer.from(saved.signSk || '', 'hex');
+    info.signPk = Buffer.from(saved.signPk || '', 'hex');
     
     info.pairedClients = {};
     for (var username in saved.pairedClients || {}) {
       var publicKey = saved.pairedClients[username];
-      info.pairedClients[username] = new Buffer(publicKey, 'hex');
+      info.pairedClients[username] = Buffer.from(publicKey, 'hex');
     }
     
     info.configVersion = saved.configVersion || 1;
