@@ -27,7 +27,7 @@ var Chacha20NonceSize =  8;
 function Chacha20Ctx() {
     this.input = new Array(16);
     this.leftover = 0;
-    this.buffer = new Buffer(Array(64));
+    this.buffer = Buffer.alloc(64);
 };
 
 function load32(x, i) {
@@ -153,7 +153,7 @@ function chacha20_update(ctx, dst, src, inlen) {
         if (src.length > 0) {
             src.copy(ctx.buffer,ctx.leftover,0,src.length);
         } else {
-            var zeros = Buffer(Array(inlen));
+            var zeros = Buffer.alloc(inlen);
             zeros.copy(ctx.buffer,ctx.leftover,0,inlen);
         }
         ctx.leftover += inlen;
