@@ -108,10 +108,10 @@ function EventedHTTPServerConnection(clientSocket) {
   this.sessionID = uuid.generate(clientSocket.remoteAddress + ':' + clientSocket.remotePort);
 
   this._remoteAddress = clientSocket.remoteAddress; // cache because it becomes undefined in 'onClientSocketClose'
-  this._pendingClientSocketData = Buffer(0); // data received from client before HTTP proxy is fully setup
+  this._pendingClientSocketData = Buffer.alloc(0); // data received from client before HTTP proxy is fully setup
   this._fullySetup = false; // true when we are finished establishing connections
   this._writingResponse = false; // true while we are composing an HTTP response (so events can wait)
-  this._pendingEventData = Buffer(0); // event data waiting to be sent until after an in-progress HTTP response is being written
+  this._pendingEventData = Buffer.alloc(0); // event data waiting to be sent until after an in-progress HTTP response is being written
   
   // clientSocket is the socket connected to the actual iOS device
   this._clientSocket = clientSocket;
@@ -159,14 +159,14 @@ EventedHTTPServerConnection.prototype.sendEvent = function(event, data, contentT
   debug("[%s] Sending HTTP event '%s' with data: %s", this._remoteAddress, event, data.toString('utf8'));
   
   // ensure data is a Buffer
-  data = Buffer(data);
+  data = Buffer.from(data);
   
   // format this payload as an HTTP response
-  var linebreak = new Buffer("0D0A","hex");
+  var linebreak = Buffer.from("0D0A","hex");
   data = Buffer.concat([
-    new Buffer('EVENT/1.0 200 OK'), linebreak,
-    new Buffer('Content-Type: ' + contentType), linebreak,
-    new Buffer('Content-Length: ' + data.length), linebreak,
+    Buffer.from('EVENT/1.0 200 OK'), linebreak,
+    Buffer.from('Content-Type: ' + contentType), linebreak,
+    Buffer.from('Content-Length: ' + data.length), linebreak,
     linebreak,
     data
   ]);
@@ -193,7 +193,7 @@ EventedHTTPServerConnection.prototype._sendPendingEvents = function() {
   }
   
   // clear the buffer
-  this._pendingEventData = new Buffer(0);
+  this._pendingEventData = Buffer.alloc(0);
 }
 
 // Called only once right after constructor finishes
