@@ -14,11 +14,11 @@ function HKDF(hashAlg, salt, ikm, info, size) {
   hmac.update(ikm);
   var prk = hmac.digest();
 
-  var prev = new Buffer(0);
+  var prev = Buffer.alloc(0);
   var output;
   var buffers = [];
   var num_blocks = Math.ceil(size / hashLength);
-  info = new Buffer(info);
+  info = Buffer.from(info);
 
   for (var i=0; i<num_blocks; i++) {
     var hmac = crypto.createHmac(hashAlg, prk);
@@ -26,7 +26,7 @@ function HKDF(hashAlg, salt, ikm, info, size) {
     var input = Buffer.concat([
       prev,
       info,
-      new Buffer(String.fromCharCode(i + 1))
+      Buffer.from(String.fromCharCode(i + 1))
     ]);
     hmac.update(input);
     prev = hmac.digest();
