@@ -29,7 +29,7 @@ function fromHex(h) {
 }
 
 function generateCurve25519SecretKey() {
-  var secretKey = Buffer(32);
+  var secretKey = Buffer.alloc(32);
   curve25519.makeSecretKey(secretKey);
   return secretKey;
 }
@@ -47,18 +47,18 @@ function generateCurve25519SharedSecKey(secKey, pubKey) {
 //Security Layer Enc/Dec
 
 function layerEncrypt(data, count, key) {
-  var result = Buffer(0);
+  var result = Buffer.alloc(0);
   var total = data.length;
   for (var offset = 0; offset < total; ) {
     var length = Math.min(total - offset, 0x400);
-    var leLength = Buffer(2);
+    var leLength = Buffer.alloc(2);
     leLength.writeUInt16LE(length,0);
 
-    var nonce = Buffer(8);
+    var nonce = Buffer.alloc(8);
     writeUInt64LE(count.value++, nonce, 0);
 
-    var result_Buffer = Buffer(Array(length));
-    var result_mac = Buffer(Array(16));
+    var result_Buffer = Buffer.alloc(length);
+    var result_mac = Buffer.alloc(16);
     encryptAndSeal(key, nonce, data.slice(offset, offset + length),
       leLength,result_Buffer, result_mac);
 
@@ -70,16 +70,16 @@ function layerEncrypt(data, count, key) {
 }
 
 function layerDecrypt(packet, count, key) {
-  var result = Buffer(0);
+  var result = Buffer.alloc(0);
   var total = packet.length;
 
   for (var offset = 0; offset < total;) {
     var realDataLength = packet.slice(offset,offset+2).readUInt16LE(0);
 
-    var nonce = Buffer(8);
+    var nonce = Buffer.alloc(8);
     writeUInt64LE(count.value++, nonce, 0);
 
-    var result_Buffer = Buffer(Array(realDataLength));
+    var result_Buffer = Buffer.alloc(realDataLength);
 
     if (verifyAndDecrypt(key, nonce, packet.slice(offset + 2, offset + 2 + realDataLength),
       packet.slice(offset + 2 + realDataLength, offset + 2 + realDataLength + 16),
@@ -100,8 +100,8 @@ function verifyAndDecrypt(key,nonce,ciphertext,mac,addData,plaintext) {
   var ctx = new chacha20poly1305.Chacha20Ctx();
   chacha20poly1305.chacha20_keysetup(ctx, key);
   chacha20poly1305.chacha20_ivsetup(ctx, nonce);
-  var poly1305key = Buffer(64);
-  var zeros = Buffer(Array(64));
+  var poly1305key = Buffer.alloc(64);
+  var zeros = Buffer.alloc(64);
   chacha20poly1305.chacha20_update(ctx,poly1305key,zeros,zeros.length);
 
   var poly1305_contxt = new chacha20poly1305.Poly1305Ctx();
@@ -112,20 +112,20 @@ function verifyAndDecrypt(key,nonce,ciphertext,mac,addData,plaintext) {
     addDataLength = addData.length;
     chacha20poly1305.poly1305_update(poly1305_contxt, addData, addData.length);
     if ((addData.length % 16) != 0) {
-      chacha20poly1305.poly1305_update(poly1305_contxt, Buffer(Array(16-(addData.length%16))), 16-(addData.length%16));
+      chacha20poly1305.poly1305_update(poly1305_contxt, Buffer.alloc(16-(addData.length%16)), 16-(addData.length%16));
     }
   }
 
   chacha20poly1305.poly1305_update(poly1305_contxt, ciphertext, ciphertext.length);
   if ((ciphertext.length % 16) != 0) {
-    chacha20poly1305.poly1305_update(poly1305_contxt, Buffer(Array(16-(ciphertext.length%16))), 16-(ciphertext.length%16));
+    chacha20poly1305.poly1305_update(poly1305_contxt, Buffer.alloc(16-(ciphertext.length%16)), 16-(ciphertext.length%16));
   }
 
-  var leAddDataLen = Buffer(8);
+  var leAddDataLen = Buffer.alloc(8);
   writeUInt64LE(addDataLength, leAddDataLen, 0);
   chacha20poly1305.poly1305_update(poly1305_contxt, leAddDataLen, 8);
 
-  var leTextDataLen = Buffer(8);
+  var leTextDataLen = Buffer.alloc(8);
   writeUInt64LE(ciphertext.length, leTextDataLen, 0);
   chacha20poly1305.poly1305_update(poly1305_contxt, leTextDataLen, 8);
 
@@ -146,8 +146,8 @@ function encryptAndSeal(key,nonce,plaintext,addData,ciphertext,mac) {
   var ctx = new chacha20poly1305.Chacha20Ctx();
   chacha20poly1305.chacha20_keysetup(ctx, key);
   chacha20poly1305.chacha20_ivsetup(ctx, nonce);
-  var poly1305key = Buffer(64);
-  var zeros = Buffer(Array(64));
+  var poly1305key = Buffer.alloc(64);
+  var zeros = Buffer.alloc(64);
   chacha20poly1305.chacha20_update(ctx,poly1305key,zeros,zeros.length);
 
   var written = chacha20poly1305.chacha20_update(ctx,ciphertext,plaintext,plaintext.length);
@@ -161,20 +161,20 @@ function encryptAndSeal(key,nonce,plaintext,addData,ciphertext,mac) {
     addDataLength = addData.length;
     chacha20poly1305.poly1305_update(poly1305_contxt, addData, addData.length);
     if ((addData.length % 16) != 0) {
-      chacha20poly1305.poly1305_update(poly1305_contxt, Buffer(Array(16-(addData.length%16))), 16-(addData.length%16));
+      chacha20poly1305.poly1305_update(poly1305_contxt, Buffer.alloc(16-(addData.length%16)), 16-(addData.length%16));
     }
   }
 
   chacha20poly1305.poly1305_update(poly1305_contxt, ciphertext, ciphertext.length);
   if ((ciphertext.length % 16) != 0) {
-    chacha20poly1305.poly1305_update(poly1305_contxt, Buffer(Array(16-(ciphertext.length%16))), 16-(ciphertext.length%16));
+    chacha20poly1305.poly1305_update(poly1305_contxt, Buffer.alloc(16-(ciphertext.length%16)), 16-(ciphertext.length%16));
   }
 
-  var leAddDataLen = Buffer(8);
+  var leAddDataLen = Buffer.alloc(8);
   writeUInt64LE(addDataLength, leAddDataLen, 0);
   chacha20poly1305.poly1305_update(poly1305_contxt, leAddDataLen, 8);
 
-  var leTextDataLen = Buffer(8);
+  var leTextDataLen = Buffer.alloc(8);
   writeUInt64LE(ciphertext.length, leTextDataLen, 0);
   chacha20poly1305.poly1305_update(poly1305_contxt, leTextDataLen, 8);
 
