@@ -11,28 +11,28 @@ module.exports = {
 
 function encode(type, data /*, type, data, type, data... */) {
     
-    var encodedTLVBuffer = Buffer(0);
+    var encodedTLVBuffer = Buffer.alloc(0);
 
     // coerce data to Buffer if needed
     if (typeof data === 'number')
-      data = Buffer([data]);
+      data = Buffer.from([data]);
     else if (typeof data === 'string')
-      data = Buffer(data);
+      data = Buffer.from(data);
     
     if (data.length <= 255) {
-        encodedTLVBuffer = Buffer.concat([Buffer([type,data.length]),data]);
+        encodedTLVBuffer = Buffer.concat([Buffer.from([type,data.length]),data]);
     } else {
         var leftLength = data.length;
-        var tempBuffer = Buffer(0);
+        var tempBuffer = Buffer.alloc(0);
         var currentStart = 0;
         
         for (; leftLength > 0;) {
             if (leftLength >= 255) {
-                tempBuffer = Buffer.concat([tempBuffer,Buffer([type,0xFF]),data.slice(currentStart, currentStart + 255)]);
+                tempBuffer = Buffer.concat([tempBuffer,Buffer.from([type,0xFF]),data.slice(currentStart, currentStart + 255)]);
                 leftLength -= 255;
                 currentStart = currentStart + 255;
             } else {
-                tempBuffer = Buffer.concat([tempBuffer,Buffer([type,leftLength]),data.slice(currentStart, currentStart + leftLength)]);
+                tempBuffer = Buffer.concat([tempBuffer,Buffer.from([type,leftLength]),data.slice(currentStart, currentStart + leftLength)]);
                 leftLength -= leftLength;
             }
         };
